Added Rainbow Feature
added changed colors
