+++ src/application/Pulser.fxml
@@ -1,0 +1,70 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import java.lang.*?>
+<?import javafx.scene.control.*?>
+<?import javafx.scene.layout.*?>
+<?import javafx.scene.shape.*?>
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.control.Menu?>
+<?import javafx.scene.control.MenuBar?>
+<?import javafx.scene.control.MenuItem?>
+<?import javafx.scene.control.RadioMenuItem?>
+<?import javafx.scene.control.ToggleGroup?>
+<?import javafx.scene.layout.AnchorPane?>
+<?import javafx.scene.shape.Rectangle?>
+
+<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="600.0" prefWidth="400.0" style="-fx-background-color: black;" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="application.Controller">
+   <children>
+      <MenuBar layoutX="94.0" layoutY="62.0" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="0.0">
+        <menus>
+          <Menu mnemonicParsing="false" text="File">
+            <items>
+              <MenuItem mnemonicParsing="false" onAction="#onClose" text="Close" />
+            </items>
+          </Menu>
+          <Menu mnemonicParsing="false" text="Colors">
+            <items>
+                  <RadioMenuItem fx:id="yellowColor" mnemonicParsing="false" selected="true" text="Yellow">
+                     <toggleGroup>
+                        <ToggleGroup fx:id="colours" />
+                     </toggleGroup>
+                  </RadioMenuItem>
+                  <RadioMenuItem fx:id="blueColor" mnemonicParsing="false" text="Blue" toggleGroup="$colours" />
+                  <RadioMenuItem fx:id="redColor" mnemonicParsing="false" text="Red" toggleGroup="$colours" />
+                  <RadioMenuItem fx:id="greenColor" mnemonicParsing="false" text="Green" toggleGroup="$colours" />
+                  <RadioMenuItem fx:id="pinkColor" mnemonicParsing="false" text="Pink" toggleGroup="$colours" />
+                  <RadioMenuItem fx:id="rainbowColor" mnemonicParsing="false" text="Rainbow!" toggleGroup="$colours" />
+            </items>
+          </Menu>
+            <Menu mnemonicParsing="false" text="Songs">
+               <items>
+                  <RadioMenuItem fx:id="chronosBtn" mnemonicParsing="false" selected="true" text="Chronos">
+                     <toggleGroup>
+                        <ToggleGroup fx:id="songs" />
+                     </toggleGroup>
+                  </RadioMenuItem>
+                  <RadioMenuItem fx:id="longLiveBtn" mnemonicParsing="false" text="Long Live the New Fresh" toggleGroup="$songs" />
+               </items>
+            </Menu>
+        </menus>
+      </MenuBar>
+      <Rectangle fx:id="b1" arcHeight="5.0" arcWidth="5.0" height="50.0" layoutX="60.0" layoutY="120.0" smooth="false" stroke="WHITE" strokeType="INSIDE" width="50.0" />
+      <Rectangle fx:id="b2" arcHeight="5.0" arcWidth="5.0" height="50.0" layoutX="140.0" layoutY="120.0" smooth="false" stroke="WHITE" strokeType="INSIDE" width="50.0" />
+      <Rectangle fx:id="b3" arcHeight="5.0" arcWidth="5.0" height="50.0" layoutX="220.0" layoutY="120.0" smooth="false" stroke="WHITE" strokeType="INSIDE" width="50.0" />
+      <Rectangle fx:id="b4" arcHeight="5.0" arcWidth="5.0" height="50.0" layoutX="300.0" layoutY="120.0" smooth="false" stroke="WHITE" strokeType="INSIDE" width="50.0" />
+      <Rectangle fx:id="b5" arcHeight="5.0" arcWidth="5.0" height="50.0" layoutX="60.0" layoutY="200.0" smooth="false" stroke="WHITE" strokeType="INSIDE" width="50.0" />
+      <Rectangle fx:id="b6" arcHeight="5.0" arcWidth="5.0" height="50.0" layoutX="140.0" layoutY="200.0" smooth="false" stroke="WHITE" strokeType="INSIDE" width="50.0" />
+      <Rectangle fx:id="b7" arcHeight="5.0" arcWidth="5.0" height="50.0" layoutX="220.0" layoutY="200.0" smooth="false" stroke="WHITE" strokeType="INSIDE" width="50.0" />
+      <Rectangle fx:id="b8" arcHeight="5.0" arcWidth="5.0" height="50.0" layoutX="300.0" layoutY="200.0" smooth="false" stroke="WHITE" strokeType="INSIDE" width="50.0" />
+      <Rectangle fx:id="b9" arcHeight="5.0" arcWidth="5.0" height="50.0" layoutX="60.0" layoutY="280.0" smooth="false" stroke="WHITE" strokeType="INSIDE" width="50.0" />
+      <Rectangle fx:id="b10" arcHeight="5.0" arcWidth="5.0" height="50.0" layoutX="140.0" layoutY="280.0" smooth="false" stroke="WHITE" strokeType="INSIDE" width="50.0" />
+      <Rectangle fx:id="b11" arcHeight="5.0" arcWidth="5.0" height="50.0" layoutX="220.0" layoutY="280.0" smooth="false" stroke="WHITE" strokeType="INSIDE" width="50.0" />
+      <Rectangle fx:id="b12" arcHeight="5.0" arcWidth="5.0" height="50.0" layoutX="300.0" layoutY="280.0" smooth="false" stroke="WHITE" strokeType="INSIDE" width="50.0" />
+      <Rectangle fx:id="b13" arcHeight="5.0" arcWidth="5.0" height="50.0" layoutX="60.0" layoutY="360.0" smooth="false" stroke="WHITE" strokeType="INSIDE" width="50.0" />
+      <Rectangle fx:id="b14" arcHeight="5.0" arcWidth="5.0" height="50.0" layoutX="140.0" layoutY="360.0" smooth="false" stroke="WHITE" strokeType="INSIDE" width="50.0" />
+      <Rectangle fx:id="b15" arcHeight="5.0" arcWidth="5.0" height="50.0" layoutX="220.0" layoutY="360.0" smooth="false" stroke="WHITE" strokeType="INSIDE" width="50.0" />
+      <Rectangle fx:id="b16" arcHeight="5.0" arcWidth="5.0" height="50.0" layoutX="300.0" layoutY="360.0" smooth="false" stroke="WHITE" strokeType="INSIDE" width="50.0" />
+      <Button fx:id="btn1" defaultButton="true" layoutX="166.0" layoutY="478.0" mnemonicParsing="false" onAction="#makeTime" text="Play Song" />
+      <Button fx:id="btn2" disable="true" layoutX="182.0" layoutY="527.0" mnemonicParsing="false" onAction="#interrupt" text="End" />
+   </children>
+</AnchorPane>
